refactor(tools): move jira create-ticket-from-task route under /tools
